+++ packages/api-graphql/src/index.ts
@@ -1,7 +1,12 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import * as events from './internals/events';
+
+export { events };
+
 export { GraphQLAPI, GraphQLAPIClass, graphqlOperation } from './GraphQLAPI';
 export * from './types';
 
 export { CONNECTION_STATE_CHANGE } from './Providers/constants';
+export * from './internals/events/types';