+++ packages/notifications/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.56](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/notifications@2.0.55...@aws-amplify/notifications@2.0.56) (2024-10-31)
+
+**Note:** Version bump only for package @aws-amplify/notifications
+
+## [2.0.55](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/notifications@2.0.54...@aws-amplify/notifications@2.0.55) (2024-10-29)
+
+**Note:** Version bump only for package @aws-amplify/notifications
+
+## [2.0.54](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/notifications@2.0.53...@aws-amplify/notifications@2.0.54) (2024-10-25)
+
+**Note:** Version bump only for package @aws-amplify/notifications
+
 ## [2.0.53](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/notifications@2.0.52...@aws-amplify/notifications@2.0.53) (2024-10-21)
 
 **Note:** Version bump only for package @aws-amplify/notifications